Refactor/subtheme page (#69)
refactor: removed use client on pages and changed the height of the highlight div
